Moved everything involving the Symbol enum into their own file to clear up the main frequencies/mod.rs file
